Simpler data types; header and pit info; refactored options

+++ src/note.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include "pit.h"
 
-void pit_note_delete(ulong id, PTask pt)
+void pit_note_delete(int id, PTask pt)
 {
     
 }
